cast F77_INT to int in error messages, release candidate 3.2.91

+++ src/sl_ab13bd.cc
@@ -122,7 +122,7 @@ For internal use only.")
             error ("lti: norm: __sl_ab13bd__: exception in SLICOT subroutine AB13BD");
             
         if (info != 0)
-            error ("lti: norm: __sl_ab13bd__: AB13BD returned info = %d", info);
+            error ("lti: norm: __sl_ab13bd__: AB13BD returned info = %d", static_cast<int> (info));
 
         if (iwarn != 0)
             warning ("lti: norm: __sl_ab13bd__: AB13BD returned iwarn = %d", iwarn);
